back to prev commit

+++ server/controllers/teacher/course.controller.js
@@ -7,6 +7,12 @@ export const addCourse = async (req, res) => {
 
         const userField = role === "teacher" ? "teacherId" : "adminId";
 
+        if (!list?.length)
+            return res.json({
+                success: false,
+                message: "Courses list is empty!"
+            });
+
         await turso.execute(
             `
         DELETE FROM teacher_courses WHERE ${userField} = ?
@@ -14,14 +20,10 @@ export const addCourse = async (req, res) => {
             [userId]
         );
 
-        if (!list?.length)
-            return res.json({
-                success: true,
-                message: "Courses list set to empty"
-            });
-
+        // Build placeholders
         const placeholders = list.map(() => "(?, ?, ?, ?)").join(", ");
 
+        // Flatten values
         const values = list.flatMap(item => [
             userId,
             item.year,
